implementing pagination currently for api only

+++ app/src/main/java/com/basemosama/movies/data/network/PagedResponse.kt
@@ -3,10 +3,10 @@ package com.basemosama.movies.data.network
 import com.google.gson.annotations.SerializedName
 
 class PagedResponse<T> (
-    val page:Long,
+    val page:Int,
     val results:List<T>?,
     @SerializedName("total_pages")
-    val totalPages :Long,
+    val totalPages :Int,
     @SerializedName("total_results")
     val totalResults :Long
 )
